Bugfix: RPC/blockchain: Use int32 for maxmempool param range checking
Since we immediately multiply by 1000000 without overflow checking, and even 2^31 is unreasonably large, just use int32_t for the range check, and immediately cast to int64 as before

+++ src/rpc/mempool.cpp
@@ -391,7 +391,7 @@ static RPCHelpMan maxmempool()
                 },
         [&](const RPCHelpMan& self, const JSONRPCRequest& request) -> UniValue
 {
-    int64_t nSize = request.params[0].getInt<int64_t>();
+    int64_t nSize = request.params[0].getInt<int32_t>();
     int64_t nMempoolSizeMax = nSize * 1000000;
 
     CTxMemPool& mempool = EnsureAnyMemPool(request.context);
